mobius-skills/provider-roster-credentialing/requirements.txt
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
google-cloud-bigquery>=3.11.0
google-maps-addressvalidation>=0.3.16
httpx>=0.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
markdown>=3.4.0
weasyprint>=60.0
